Add peak memory tracking to sql-integration
