+++ Cargo.toml
@@ -1,9 +1,10 @@
 workspace = { members = [
+    "crates/polytune-test-utils",
+    "examples/api-integration",
     "examples/http-multi-server-channels",
     "examples/http-single-server-channels",
     "examples/wasm-http-channels",
     "examples/sql-integration",
-    "examples/api-integration",
 ] }
 [package]
 name = "polytune"
@@ -77,6 +78,7 @@ cpufeatures = "0.2.17"
 
 [dev-dependencies]
 criterion = { version = "0.6.0", features = ["async_tokio"] }
+polytune-test-utils = { path = "crates/polytune-test-utils" }
 proptest = "1.7.0"
 tokio = { version = "1.46.1", features = ["full"] }
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
