+++ crates/polytune-test-utils/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "polytune-test-utils"
+version = "0.1.0"
+edition = "2024"
+publish = false
+
+[dependencies]
+tokio = {version = "1.46.1", features = ["rt-multi-thread"]}
