+++ examples/sql-integration/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "polytune-sql-integration"
-version = "0.1.0"
 edition = "2024"
+name = "polytune-sql-integration"
 publish = false
+version = "0.1.0"
 
 [dependencies]
 anyhow = "1.0.79"
@@ -10,6 +10,7 @@ axum = "0.7.4"
 blake3 = "1.5.1"
 clap = { version = "4.5.41", features = ["derive"] }
 polytune = { path = "../../", version = "0.2.0-alpha.1" }
+polytune-test-utils = { path = "../../crates/polytune-test-utils" }
 reqwest = { version = "0.12.22", features = ["json"] }
 serde = "1.0.197"
 serde_json = "1.0.115"
